chore: mypy strict on domains/platform/cli (P3)
mypy>=1.13 dev dep；[tool.mypy] strict + pydantic.mypy plugin。
overrides: app.* / integrations.* 暂时放宽 disallow_untyped_defs/incomplete_defs/
untyped_decorators/disallow_any_generics/warn_return_any 与若干 error code，
其余 strict 项仍全局生效；override 名单写入 TECH_DEBT 跟踪。
pre-commit mypy hook 启用；make check 加 mypy 行。

domains/ 修复 162 → 0：
- 大量 dict | None 类型窄化（_dict_or_empty 模式 + 显式 isinstance 重绑定）
- runtime worker 注入参数补 Protocol/Any 注解（依赖反向；保持架构边界）
- async Task[None] 泛型补全
- 多处 Returning Any 通过本地类型注解或显式 cast 收紧
- eth_utils 改用 .address 子模块导入以满足 strict re-export 检查

platform/ 修复 5 → 0；cli.py 已是 strict-clean。

Real bug fixed: catalyst_ranking_service.py 的 `from .discussion_quality_scoring`
是失败导入（模块在 scoring/ 包），改为绝对路径并加 regression test
tests/unit/test_catalyst_ranking_service_imports.py 锁住。

Owning spec: docs/superpowers/specs/active/2026-05-10-tests-and-lint-production-grade-design-cn.md
Phase: 3/6

+++ src/gmgn_twitter_intel/domains/token_intel/read_models/catalyst_ranking_service.py
@@ -4,7 +4,7 @@ import json
 import math
 from typing import Any
 
-from .discussion_quality_scoring import post_quality_score
+from gmgn_twitter_intel.domains.token_intel.scoring.discussion_quality_scoring import post_quality_score
 
 CATALYST_OBSERVATION_MS = 30 * 60_000
 CATALYST_BASELINE_MS = 60 * 60_000
@@ -150,7 +150,7 @@ def _avg_followup_quality(followups: list[dict[str, Any]]) -> float:
         / 100
         for row in followups
     ]
-    return sum(scores) / len(scores)
+    return float(sum(scores) / len(scores))
 
 
 def _reference(value: Any) -> dict[str, Any] | None:
